src/requirements.txt
Django
djangorestframework
black
isort
